Disable factory-reset feature and remove files associated with it
This feature needs more integration work first before it can be enabled
in Debian. E.g. Debian still uses a split /usr setup and a merged /usr is
a prerequisite for factory-reset to work properly.
Atm. it also leads to dependency cycles between keyboard-setup.service
and systemd-udev-hwdb-update.service.

This should be revisited for jessie+1.