Update change log; make eudev fork official.
Add epoch version equal to the systemd release on which this
eudev is based.  This is technically udev's version anyway as
systemd abandoned its own version when it merged with udev.

Since eudev restarted its numbering, eudev's packages are
considered too old (1.9 <<<< 200) to be used with just about
any package that requires udev support, which is why the epoch
version is being used.